Authored LILYPOND-UPDATES document describing how to update test artifacts.
Ten commands (a mixture of IDE commands and shell commands) appear at the head
of the document in the summary section.

Run these 10 commands after downloading a new version of the LilyPond.

The remainder of the (verbose) document records the procedure by which these
commands were determined. The method section of the document can be used to
rederive a corresponding set of commands should IDE behavior change in future.