Implemented graphing protocol via __graph__() special method. MODEL.
MODEL: This commit refactors the codebase to implement a graphing protocol
analogous to the current illustration protocol.

All classes which previously implemented a `graphviz_format` property or a
`graphviz_graph` property now simply implement a `__graph__()` special method.
This method returns a GraphvizGraph instance. The `topleveltools.graph()`
function, like `show()` now looks for that method when attempting to generate a
Graphviz graphh of an object.