Added new Transition class to indicatortools.
The class is current a stub.

The class encapsulate the graphic properties of the line-drawn part of text
spanners: line dashing, endpoint padding, endpoint arrows and so on.

This new Transition class will be used together with the existing Markup and
TextSpanner classes to afford multistage text spanner definition in a piecewise
way.