Extended CountsSelectorCallback.
You can now use negative counts. They will be treated as positive for the
purposes of partitioning selections, but the partitioned parts associated with
the negative counts will not be included in the result.

This is useful for creating selected and un-selected runs of leaves. The
following selector will skip 7 leaves, select 2, skip 3, then select 1:

    selectortools.Selector(
        ).by_leaves(
        ).by_counts(
            [-7, 2, -3, 1],
            )