Improved LilyPondFormatManager.
When formatting non-scoped indicators, LilyPondFormatManager will attempt to
find and use the indicator's `_get_lilypond_format_bundle()` method before it
tries to get a bundle via the indicator's `_lilypond_format_bundle` property.

The method `_get_lilypond_format_bundle()` is conceptually identical to the
identically named method on `Spanner`. It allows for providing format
contributions conditionally based on the component to which the indicator is
attached.