Fixed interaction between TextSpanner and annotated markup.
Consider this example:

    staff = Staff("c'4 d'4 e'4 f'4")
    markup = Markup('foo')
    attach(markup, staff[0], is_annotation=True)
    attach(spannertools.TextSpanner(), staff[:])

OLD OUTPUT:

    >>> f(staff)
    \new Staff {
        c'4 ^ \markup { foo }
        d'4
        e'4
        f'4 \stopTextSpan
    }

NEW OUTPUT:

    >>> f(staff)
    \new Staff {
        c'4 ^ \markup { foo }
        d'4
        e'4
        f'4
    }

Explanation: First note that 'annotated markup' means markup attached to a
component with is_annotation=True set in the call to attach(). Then note that
2.16 greatly extends the behavior of TextSpanner (which should be included in
the 2.16 release notes). TextSpanner now interacts with annotated markup in
intelligent ways. It remains to document the complete set of new behaviors that
determine the ways that TextSpanner interacts with annotated markup. But
suffice it to say for this commit that TextSpanner now knows about annotated
markup and also about the new LineSegment and Arrow classes. When
TextSpanner find Markup, LineSegment or Arrow objects attached to leaves,
TextSpanner formats all of these objects together in an intelligent way. This
sort of group formatting sometimes includes silencing the \startTextSpan and
\stopTextSpan commands that are the usual output of TextSpanner. The bug fixed
in this commit causes a spurious \stopTextSpan to no longer appear.