Changed experimental.makers.PianoStaffSegmentMaker.__call__() return value.
OLD: Returned LilyPond file.

NEW: Returns LilyPond file and "sticky settings" dictionary together as a pair.

This represents a change in the protocol that all segment-makers are expected
to follow: the __call__() method of all segment-makers should now return the
pair of a LilyPond file together with a sticky settings dictionary for the
segment in question.

Change motivated by work allowing the IDE to manage intersegment sticky
settings.