Improved Spanner handling in Container._set_item().
This commit fixes the following edge-case: given a container where the first
child shares spanners with the container's previous sibling, or where the last
child shares spanners with the container's next sibling, components inserted at
the beginning or end of that container will also be properly inserted into
those shared spanners.

Consider a staff container four leaves, with the inner two leaves grouped into
their own container:

    >>> staff = Staff("c'4 { d'4 e'4 } g'4")

We can attach a beam to the leaves in the staff, resulting in expected output:

    >>> beam = Beam()
    >>> attach(beam, staff.select_leaves())
    >>> print(format(staff))
    \new Staff {
        c'4 [
        {
            d'4
            e'4
        }
        f'4 ]
    }

The beam's component list also corresponds with the staff's leaves:

    >>> beam.components
    Selection(Note("c'4"), Note("d'4"), Note("e'4"), Note("f'4"))

After inserting a component into the middle of the inner container, that
component appears in the beam's component list:

    >>> rest = Rest((1, 4))
    >>> staff[1].insert(1, rest)
    >>> rest in beam.components
    True

However, previous to this commit, any component inserted at the beginning or
end of the inner container would not appear in the beam's component list:

    >>> chord_1 = Chord("<c' e' g'>4")
    >>> chord_2 = Chord("<d' f' a'>4")
    >>> staff[1].insert(0, chord_1)
    >>> staff[1].append(chord_2)
    >>> chord_1 in beam.components
    False
    >>> chord_2 in beam.components
    False
    >>> beam.components
    Selection(Note("c'4"), Note("d'4"), Rest("r4"), Note("e'4"), Note("f'4"))

After this commit, both `chord_1` and `chord_2` appear in the beam's component
list in the expected locations:

    >>> for component in beam.components:
    ...     component
    ...
    Note("c'4")
    Chord("<c' e' g'>4")
    Note("d'4")
    Rest('r4')
    Note("e'4")
    Chord("<d' f' a'>4")
    Note("f'4")

All tests pass, and docs continue to build.