Added previous_segment_metadata=None to SegmentMaker.__call__() method.
The new signature is this:

    SegmentMaker.__call__(
        self,
        segment_metadata=None,
        previous_segment_metadata=None,
        )

This is the first in a couple of commits that will cause the IDE to pass two
metadata dictionaries to every segment-maker at illustration-time: one metadata
dictionary for the current segment and a second metadata dictionary for the
previous segment.

Motivation is to allow every segment to know about end settings (like effective
instrument, clef, tempo) of the segment that ends just before.