Added check_conflicting_clefs() method to WellformednessManager. NOTE.
Conflicting clefs is a relatively obscure situation defined equal to two
different clefs being set on components that start at the same time in the same
score.

An exception raises and prevents composers from doing this directly.

But there are relatively straight-forward cases when composers can create this
situation without triggering the exception.

NOTE: An extensive docstring is included with this commit and should be read.

The check now runs automatically as part of any call to
insepct_()is_well_formed().