Improved tuplet number formatting corner case.
OLD:

    This worked:

    >>> tuplet = Tuplet((2, 3), "c'8 c' c'")
    >>> override(tuplet).tuplet_number.text = Markup('foo')

    But this did not work:

    >>> tuplet = Tuplet((3, 5), "c'8 c' c' c' c'")
    >>> override(tuplet).tuplet_number.text = Markup('foo')

NEW: Both examples now work.

Explanation: Abjad treats tuplets-that-fit-into-non-power-of-two-durations in a
way such that fraction text (like 5:3) always appears. This is good because "5"
is ambiguous and probabaly means 5:4 ... which makes printing "5" especially
bad if the tuplet ratio is actually 5:3. But this special-casing of such
tuplets was preventing the ability to override tuplet number text explicitly.
This commit extends the special-casing code and makes it possible to override
tuplet number text explicitly for the first time.

Doesn't need to be included in upcoming release notes.