Added remember_state=False keyword to all rhythm-maker __call__() methods. NEW.
The keyword is fully active in EvenDivisionRhythmMaker.

The keyword has not yet integrated into the other rhythm-makers.

Examples: rebuild the docs and see the EvenDivisionRhythmMaker.__call__
docstring before contrastive examples.

Explanation: Takes advantage of the magic state-preservation features of Python
generators. A new static helper method is implemented on RhythmMaker:

    @staticmethod
    def _make_cyclic_tuple_generator(iterable):
        cyclic_tuple = datastructuretools.CyclicTuple(iterable)
        i = 0
        while True:
            yield cyclic_tuple[i]
            i += 1

The method wraps a cyclic tuple in a generator. The method is called *at the
end of initialization*. The generator is stored on the object created at the
end of initialization. This adds only a single extra line (per iterable) to the
rhythm-maker initializer. Here're the relevant lines in
EvenDivisionRhythmMaker:

        self._denominators_generator = self._make_cyclic_tuple_generator(
            denominators)
        self._extra_counts_per_division_generator = \
            self._make_cyclic_tuple_generator(extra_counts_per_division)

Call-time behavior then checks the remember_state flag. If remember_state is
false then music-making proceeds as before. If remember_state is true then
values are obtained for the generators cached during initialization. This
amounts to two extra lines in EvenDivisionRhythmMaker:

            if remember_state:
                denominator = self._denominators_generator.next()
                extra_count = self._extra_counts_per_division_generator.next()
            else:
                denominator = denominators[i]
                extra_count = extra_counts_per_division[i]

This technique will absolutely work for all iterables in all rhythm-makers.

The question is deciding whether we want the functionality implemented or not.