Added new PartitionTable class to rhythmmakertools.
The class inherits from TypedOrderedDict.

The class provides a way for composers to specify arbitrary ratios used for
dividing temporal objects.

For example, The integer 7 can be entered in the table as 7 = 4 + 3 or
as 7 = 1 + 6. (This expresses an integer partition as a ratio.)

Some of the rhythm-makers will integrate PartitionTable to customize tuplet
creation and other parts of rhythmic output.