Activated TupletSpellingSpecifier.flatten_trivial_tuplets in rhythm-makers.
New functionality.

See TupletRhythmMaker.tuplet_spelling_specifier docstring for examples.

NOTE: the motivation here is the fact that untupletted notes can be
meter-rewritten in ways that tupletted notes can not. So there's a real-world
benefit to setting flatten_trivial_tuplets=True if you then subsequently
rewrite the meters of the rhythms produced.