Cleaned up TimeSignature initializer. CHANGE.
This commit reduces the signature polymorphism of the TimeSignature
initializer.

Which is a fancy way of saying that this commit removes *args and **kwargs from
the TimeSignature initializer.

This conceivably takes away functionality, as follows:

OLD: TimeSignature(4, 4) was allowed.

NEW: Only TimeSignature((4, 4)) is allowed.

CHANGE: should be noted in the 2.16 change log.

It is now the case that all indicators implement explicit initializers.