Changed Abjad dependency installation. NOTE.
NOTE: Python's setuptools allows you to specify both the core dependencies for
a project as well as optional auxiliary dependencies. Core dependencies, listed
via the "install_requires" keyword of setuptools.setup() are always installed
when you install a project. Additional optional dependencies can be specified
via the "extras_require" keyword, which accepts a dictionary where the keys are
string labels and the values are lists of dependencies. Prior to this commit,
Abjad treated pytest and sphinx as core dependencies, although these are only
ever used when developing the system of building the documentation. After this
commit, pytest and sphinx are only installed with Abjad when one requests
additional development dependencies.

To install Abjad normally:

    pip install abjad

To install with Sphinx and pytest:

    pip install abjad[development]

This change will not affect users who already have these dependencies
installed; nothing is removed. New users or projects using Abjad itself as a
dependency (and additionally using Sphinx or pytest) should adjust their Abjad
installation commands.

Should be noted in 2.17 release notes.