Added initial example to custom spanner cookbook article.
For some reason, I wasn't able to simply reproduce hidden versions of the
classes and functions used for the final result in order to create the teaser
image. `inspect_(leaf).get_indicators(OscillationSize)` never found any
indicators! Some kind of globals/locals problem? The workaround was to make the
annotation class an inner class of the spanner. Still, why does it all work at
the very end of the example?