Added pytest.ini. NOTE.
NOTE: py.test knows to looks for a number of configuration files, such
as setup.cfg, pytest.ini and others. You can define various commands in
these files. In this instance, the sole contents of the file are:

    [pytest]
    norecursedirs = abjad/boilerplate

This tells py.test to ignore the contents of the boilerplate directory
when collecting tests. py.test will find this ini file no matter what
directory it is called from.