Added new 'exact partitioning' functionality to partition_sequence_by_counts().
NEW.

Here're examples of the preexisting overhang functionality of
sequencetools.partition_sequence_by_counts():

    >>> sequencetools.partition_sequence_by_counts(range(10), [3], overhang=False)
    [[0, 1, 2]]

    >>> sequencetools.partition_sequence_by_counts(range(10), [3], overhang=True)
    [[0, 1, 2], [3, 4, 5, 6, 7, 8, 9]]

NEW: you can now use the Exact ordinal constant to set overhang=Exact:

    >>> sequencetools.partition_sequence_by_counts(range(10), [3], overhang=Exact)
    Traceback (most recent call last):
    File "<stdin>", line 1, in <module>
    File "/Users/trevorbaca/Documents/abjad/abjad/tools/sequencetools/partition_sequence_by_counts.py", line 217, in partition_sequence_by_counts
        raise Exception(message)
    Exception: sequence does not partition exactly.

This is good because it's a way of ensuring that your partition is exact (and
raising an exception when it is not).

Note that this behavior is correct, too, because there would be an 8-element
overhang were the function called with overhang=True:

    >>> sequencetools.partition_sequence_by_counts(range(10), [2], overhang=Exact)
    Traceback (most recent call last):
    File "<stdin>", line 1, in <module>
    File "/Users/trevorbaca/Documents/abjad/abjad/tools/sequencetools/partition_sequence_by_counts.py", line 217, in partition_sequence_by_counts
        raise Exception(message)
    Exception: sequence does not partition exactly.

And not that this works because there is no overhang when the [2] counts
parameter is interpreted cyclically:

    >>> sequencetools.partition_sequence_by_counts(range(10), [2], cyclic=True, overhang=Exact)
    [[0, 1], [2, 3], [4, 5], [6, 7], [8, 9]]

NEW. Should be included in the release notes.