Added MultipartBeam.beam_rests property.
When MultipartBeam.beam_rests is set to false rests are not included in the
beam.

When MultipartBeam.beam_rests is set to true rests are included in the beam.

The property defaults to false.

This commit implements new functionality. But because an overhaul of the beam
spanners is still pending, this commit does not need to be included in release
notes.