Wrote selector test demonstrating arbitrary object selection. NEW.
Consider the following example:

    >>> selector = selectortools.Selector()
    >>> selector = selector.by_counts([3])
    >>> numbers = [1, 'two', 'three', 4, -5, 'foo', 7.0, 8]
    >>> selector(numbers)
    Selection(Selection(1, 'two', 'three'),)

The selector callback classes were designed to select components.

But note that the selector callback classes can be used to select arbitrary
objects, too.

This is important. This means that the selector callback classes can be created
and configured as input to *arbitrary pieces of code* in the system.

Stated even more generally, the "select(or) protocol" can be used throughout
the entire system.

For example, consider a yet-to-be developed subsystem designed to select (time)
"divisions" in interesting ways. Just a division selection system can and
should be implemented in terms of the existing selector classes.

NEW. No new code included in this commit. Rather the new test included in this
commit is a type of proof-of-concept that the existing selector classes can be
used to select arbitrary objects (in addition to the existing role of selecting
components).