Ported three tests forward for Python 3 compatability.
Note that all system tests were passing for Python 2 and Python 3, with only
the exception of the Python 3.5.

So the three tests changed in this commit were changed to accommodate something
in Python 3.5 that is not present in Python 2.x, 3.3 or 3.4.

Two of the three tests concerned function call counts against Fraction and
NonreducedFraction objects. (So Python 3.5 initializes Fraction objects within
fewer function calls than do Pythons 3.3 and 3.4.)

The third test concerned Python's built-in "inspect" module. Python 3.5
evidently changed something in the public interface to "inspect" that differs
from Pythons 3.3 and 3.4.