Taught selectors about start offsets. NEW.
Consider the following five notes. We will call these an 'input selection':

    >>> notes = scoretools.make_notes([0, 2, 4, 5, 7], [(1, 4)])
    >>> notes
    Selection(Note("c'4"), Note("d'4"), Note("e'4"), Note("f'4"), Note("g'4"))

We can use selectors to select the last two notes. We will call the result an
'output selection':

    >>> selector = selectortools.Selector()
    >>> selector = selector.get_slice(start=-2)
    >>> selector(notes)
    Selection(Selection(Note("f'4"), Note("g'4")),)

QUESTION: what is start offset of the input selection?

QUESTION: what is the start offset of the output selection?

This commit extends selectors to answer these questions.

First, note that no start offset is given for the input selection. (It probably
makes sense to think of the start offset of the input selection as equal to
zero. But this might sometimes be otherwise. For example, when building up
chunks of music to place at arbitrary locations in a score, analogous to
thinking in terms of timespans.)

Second, note that the start offset of the output selection occurs 3/4 after the
start offset of the input selection. (This is easy to see by eyeballing.)

Selectors now model all of this with the new start_offset keyword passed in at
call-time:

    >>> selector(notes, start_offset=0)
    (Selection(Selection(Note("f'4"), Note("g'4")),), Offset(3, 4))

Old behavior continues as before: just leave start_offset set to none.

New behavior occurs when start_offset is set to something other than none. In
this case the selector returns a selection of selections (just like before)
together with the start offset of the (first) output selection.

Note that the start offset of the output selection is measured from the value
of the start_offset keyword (which is taken to model the start offset of the
input selection).

NOTE. This commit changes the signature all selector callback __call__ methods.

NOTE. This commit implements start offset handling in SliceSelectorCallback
only. Other selectors will be taught about start offset handling in future
commits.