Refactored Division.
OLD: Division aggregated NonreducedFraction.
NEW: Division inherits from NonreducedFraction.

EXPLANATION. Aggregation is frequently better than inheritance. But Division
models a portion of the metric hierarachy. This means that Division((6, 4))
differs from Division((3, 2)). The old implementation of Division was lossy on
initialization and converted (6, 4) to (3, 2). This new implementation of
Division is not lossy and preserves the difference between (6, 4) and (3, 2).