Deprecated Selector start_offset evaluation keyword.
OLD:

    Selector.__call__(self, expr, rotation=None, start_offset=None)

NEW:

    Selector.__call__(self, expr, rotation=None)

This commit doesn't yet implement the new signature. This commit just announces
that the new signature is pending in a commit that is soon to arrive.

EXPLANATION. The ideal signature for callback evaluation would be ...

    Selector.__call__(self, expr)

... or even ...

    Selector.__call__(self, *expr)

... with the named keywords removed (and rotation effected through an explicit
callback). But the rotation=None keyword is of course 100% ok; it represents an
established compositional practice already deployed in multiple scores.

Removing the start_offset=None keyword in a pending commit means that the
application of the callbacks in the selector callback stack can proceed much
more directly.