Added new expressiontools package. NEW.
Contents:

    * expressiontools.Callback
    * expressiontools.SequenceExpression
    * expressiontools.sequence

Package can grow like this in future:

    * expressiontools.Callback
    * expressiontools.SequenceExpression
    * expressiontools.IterationExpression
    * expressiontools.iterate
    * expressiontools.sequence

OBSERVATION. A surprising amount of the work of composing music involves
iteration and sequence manipulation.

OBSERVATION. Most of the functionality implemented in selectortools are
iteration expressions.

DESIGN. It's possible that unifying sequence and iteration methods might
produce a unified system of callbacks that is comprehensive for music
composition.