Added TaleaRhythmMaker.read_talea_once_only property. NEW.
Defaults to none.

Set TaleaRhythmMaker.read_talea_once_only to true to use the input talea in a
noncyclic way. This enables composers to use the input talea as a type of fully
worked-out "pad" of values that has been precomposed ahead of time, as for
example when working out an additive ritardando with counts that go from
smaller to larger over the course of all divisions passed into the
rhythm-maker.

API includes docstring.

Pytest tests that exception raises.