Extended abjadbooktools to allow exceptions in literal code blocks. NEW.
You can now document exception behavior in the public API. The functionality is
very much the same as in doctest.

EXAMPLE: This is now OK in a docstring:

    Calling maker on these divisions raises an exception because talea
    is too short to read once only:

    ::

        >>> divisions = [(3, 8), (3, 8), (3, 8), (3, 8)]
        >>> maker(divisions)
        Traceback (most recent call last):
        ...
        Exception: talea is too short to read once only:
        CyclicTuple([1, 2, 3, 4]) in [6, 6, 6, 6].

Notice that ellipses may be used as usual to eliminate line number- and
filesystem-specific information from the docstring.

Much thanks for the guidance in GitHub #536 for this one. Worked like a charm.