Fixed bug in TextSpanner handling of annotated v. nonannotated markup handling.
OLD: there was a small bug when use the piecewise definition protocol (PDP) to
create text spanners. When using annotated markup to define the start-markup
and start-markup of a text spanner no bug arose. But when doing this together
with an independent (ie, nonannotated) markup attached to the same leaf as eith
either the start-markup or stop-markup of the spanner, the bug appeared. The
character of the bug is that the text spanner would incorectly capture the
nonannotated markup and use this markup in place of the annotated markup. This
was due to TextSpanner._get_annotations() incorrectly examining attached markup
to check for annotation status.

NEW: you can now freely mix annotated and nonannotated markup when using PDP
text spanners.

Added py.test regression test to create a contract for this behavior.