Commented-out one ineffectual line in documentationtools.
The function documentationtools.make_reference_manual_lilypond_file() contained
the following assignment:

    lilypond_file.line_width = lilypondfiletools.LilyPondDimension(6, 'in')

This was probably a typo. The assignment should be ...

    lilypond_file.paper_block.line_width = lilypondfiletools.LilyPondDimension(6, 'in')

... instead.

NOTE. This commit does not repair the assignment as shown above. Why not?
Because a number LaTeX tests in abjadbooktools break in a way that is not
completely clear. (These tests are tests that apparently test on hashed values
of PDFs and other assets produced during documentation production.)

Feel free to make the change indicated in this commit; just make sure to port
the corresponding abjadbooktools tests while doing so.