Commented out deprecated Container.select_leaves() method. CHANGE.
This closes #456.

CHANGE MANAGEMENT. This is a breaking change. This commit comments out the
Container.select_leaves() method instead of removing the method. This is to
make it easier to communicate the change to the user base, and to restore (ie,
uncomment) the method if it is necessary temporarily to do so while
communicating the change.