Added new StorageFormatSpecifier.include_abjad_namespace property. NEW.
Developers can now say things like this:

    @property
    def _storage_format_specification(self):
        from abjad.tools import systemtools
        ...
        return systemtools.StorageFormatSpecification(
            self,
            include_abjad_namespace=True, # <== NEW
            keyword_argument_names=keyword_argument_names,
            positional_argument_values=positional_argument_values,
            )

This causes the Abjad namespace to be preprended to tools package names in
storage format output:

    >>> dictionary = datastructuretools.TypedOrderedDict([
    ...     ('color', 'red'),
    ...     ('directive', Markup(r'\italic Allegretto')),
    ...     ])

    >>> print(format(dictionary))
    abjad.datastructuretools.TypedOrderedDict( # <== NEW
        [
            ('color', 'red'),
            (
                'directive',
                markuptools.Markup(
                    contents=(
                        markuptools.MarkupCommand(
                            'italic',
                            'Allegretto'
                            ),
                        ),
                    ),
                ),
            ]
        )

EXPLANATION. The purpose of the commit is to allow
datastructuretools.TypedOrderedDict to output its storage format in this new
way. (This change will be made in the next commit.)

The reason for doing this is to allow 'import abjad' at the head of all score
package __metadata.__py files.

NOTE. It probably makes very good sense to set include_abjad_namespace to true
by default for ALL classes in the system! This will involve updating dozens of
doctests. But this will then make the evaluation of storage format output quite
straightforward.

I think the new realization is that storage format output is almost always
produced for *external* storage. That is, true client code makes use of storage
format output. And it makes sense for client code to use 'import abjad' rather
than 'from abjad import *'. So it correspondingly will make sense to turn
include_abjad_namespace on by default for all Abjad classes. (Such a change
will not happen in a near term commit but would happen later.)