Extended iterate().depth_first() to always include grace notes.
The commit is similar to the previous commits that added a
with_grace_notes=False keyword to iterate().by_class() and iterate().by_leaf().

The difference here is that iterate().depth_first() now *ALWAYS* yields grace
notes (and the grace containers that house grace notes). The fact that
depth-first iteration of the score previously did not return grace notes was a
10-year-old gap in the system. This commit fills that gap.

Example:

    >>> voice = Voice("c'8 [ d'8 e'8 f'8 ]")
    >>> grace_notes = [Note("c'16"), Note("d'16")]
    >>> grace = scoretools.GraceContainer(
    ...     grace_notes,
    ...     kind='grace',
    ...     )
    >>> attach(grace, voice[1])
    >>> after_grace_notes = [Note("e'16"), Note("f'16")]
    >>> after_grace = scoretools.GraceContainer(
    ...     after_grace_notes,
    ...     kind='after')
    >>> attach(after_grace, voice[1])
    >>> show(voice)

    >>> for component in iterate(voice).depth_first():
    ...     component
    ...
    Voice("c'8 d'8 e'8 f'8")
    Note("c'8")
    Note("d'8")
    GraceContainer("c'16 d'16")
    Note("c'16")
    Note("d'16")
    GraceContainer("e'16 f'16")
    Note("e'16")
    Note("f'16")
    Note("e'8")
    Note("f'8")

User base doesn't need to be notified. Even though the behavior of DF iteration
is now changed globally in the system, users will not notice.