Equipped Markup with private _annotation variable.
Developers can use this slot to annotate markup function: 'is sketch
annotation', for example, could be ascribed to markups that are meant to appear
during composition but are meant to be stripped from the score for publication.

Right now there is no public access to this feature. Feature will remain in
system for a while; if feature is found to be useful in a general sense then
public access will be afforded through the inspection protocol.