Implemented label-iteration (functional) composition. POWERFUL.
OLD: LabelAgent labeled logical ties in client. (Or LabelAgent labeled vertical
moments in client with LabelAgent(expr).vertical_moments().)

NEW: LabelAgent labels arbitrary selections.

What does this mean?

Structurally, this means that iteration and labeling are now disentangled. You
can iterate whatever you want using iterate(). And then you can pass the output
of iterate() to label(). (A powerful way to think about this is as a type of
functional composition between iteration and labeling.)

In terms of user functionality, this means that arbitrary groups of things can
be labeled. The components of spanners can be labeled! Arbitrarily chosen
(hand-picked) selections can be labeled, too. And of course vertical moments
can labeled (as before).

Note that the ability to label spanner components is particularly powerful when
using HorizontalBracketSpanner as a type of analysis bracket. This example
labels horizontal bracket components with (Forte-ranked
transposition-inversion) set-classes:

    >>> string = "df''8 c''8 bf'8 a'8 f'4. fs'8 g'8 b'8 d''2."
    >>> voice = Voice(string)
    >>> voice.consists_commands.append('Horizontal_bracket_engraver')
    >>> selections = [voice[:4], voice[-4:]]
    >>> for selection in selections:
    ...     spanner = spannertools.HorizontalBracketSpanner()
    ...     attach(spanner, selection)
    ...
    >>> label(selections).with_set_classes()
    >>> override(voice).horizontal_bracket.staff_padding = 3
    >>> override(voice).text_script.staff_padding = 2

    >>> f(voice)
    \new Voice \with {
        \consists Horizontal_bracket_engraver
        \override HorizontalBracket.staff-padding = #3
        \override TextScript.staff-padding = #2
    } {
        df''8 \startGroup
            ^ \markup {
                \tiny
                    \line
                        {
                            "SC(4-3){0, 1, 3, 4}"
                        }
                }
        c''8
        bf'8
        a'8 \stopGroup
        f'4.
        fs'8 \startGroup
            ^ \markup {
                \tiny
                    \line
                        {
                            "SC(4-20){0, 1, 5, 8}"
                        }
                }
        g'8
        b'8
        d''2. \stopGroup
    }

The example should seem a little bit like magic because it's almost as though
Abjad were handling the set-class analysis all on its own. (See the API entry
for notation.)

Another implication of arbitrary selection labeling is that selections may be
made "ahead of time" (ie, before labeling) and effectively used as a type of
filter. This example labels pitch names (filtered by selection):

    >>> voice = Voice("df''4 c''4 f'4 fs'4 d''4 ds''4")
    >>> voice.consists_commands.append('Horizontal_bracket_engraver')
    >>> selections = [voice[:2], voice[-2:]]
    >>> for selection in selections:
    ...     spanner = spannertools.HorizontalBracketSpanner()
    ...     attach(spanner, selection)
    ...
    >>> label(selections).with_pitches()
    >>> override(voice).horizontal_bracket.staff_padding = 3
    >>> override(voice).text_script.staff_padding = 2

    >>> f(voice)
    \new Voice \with {
        \consists Horizontal_bracket_engraver
        \override HorizontalBracket.staff-padding = #3
        \override TextScript.staff-padding = #2
    } {
        df''4 \startGroup
            ^ \markup {
                \small
                    df''
                }
        c''4 \stopGroup
            ^ \markup {
                \small
                    c''
                }
        f'4
        fs'4
        d''4 \startGroup
            ^ \markup {
                \small
                    d''
                }
        ds''4 \stopGroup
            ^ \markup {
                \small
                    ds''
                }
    }

Note that only notes contained in analysis brackets are labeled; other notes
remain unchanged: the analysis brackets are effectively being used as a type of
filter.

MODEL. Note that this new affordance in label() effectively deprecates
label().vertical_moments(). But it will probably be a while before
label().vertical_moments() is removed.