Cleaned up tweaks: stage 1.
NOTE: this is the first of two commits that refactor tweaks for Abjad
3.7. The functionality in this commit is temporary, and will be
superseded by the functionality in the second commit, which will follow
in a few days.

The primary change is in the way abjad.tweak() works:

    OLD:

        articulation = abjad.Articulation(".")
        abjad.tweak(articulation).color = "#red"

    NEW:

        articulation = abjad.Articulation(".")
        abjad.tweak(articulation, r"- \tweak color #red")

There is also now an abjad.Tweak class:

    >>> articulation = abjad.Articulation(".")
    >>> abjad.tweak(articulation, r"- \tweak color #red")
    >>> articulation.tweaks[0]
    Tweak(string='- \\tweak color #red', tag=None)

But note, again, that this functionality will change again in the
pending commit.