Alphabetized all format contributions (by format site).
The lexical order of commands in LilyPond output no longer depends on
the (dynamic) order in which indicators are attached to leaves.

Alphabetized tags, and offset tags two spaces to the right in LilyPond
output.

Added abjad.lilypond(..., site_comments=False) keyword. Set to true to
debug the names and contents of every format site.

Closes #1450.